pyproject.toml
[project]
name = "backend-my-event"
version = "0.1.0"
description = ""
authors = [
    {name = "jamayapabon",email = "jeap1723@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "alembic (>=1.14.0,<2.0.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.10.5,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.2.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
